requirements.txt
matplotlib
python-dotenv
sagemaker
torch
